fixed adding new blog tests after adding token-based authentication

+++ controllers/blogs.js
@@ -11,8 +11,10 @@ blogsRouter.post('/', async (request, response, next) => {
     const body = request.body
   
     try {
+      if(request.token === null) {
+        return response.status(401).json({ error: 'token invalid or missing' })
+      }
       const user = request.user
-
       const blog = new Blog({
         title: body.title,
         author: body.author,
@@ -32,16 +34,19 @@ blogsRouter.post('/', async (request, response, next) => {
 
 blogsRouter.delete('/:id', async (request, response, next) => {
     try {
-    const blog = await Blog.findById(request.params.id)
-    const user = request.user
-    if(!(blog.user.toString() === user.id.toString())) {
-      return response.status(401).json({ error: 'token invalid' })
-    }
-    else {
-      await Blog.findByIdAndDelete(request.params.id)
-      response.status(204).end()
-    } 
-    }
+      if(request.token === null) {
+        return response.status(401).json({ error: 'token invalid or missing' })
+      }
+      const blog = await Blog.findById(request.params.id)
+      const user = request.user
+      if(!(blog.user.toString() === user.id.toString())) {
+        return response.status(401).json({ error: 'token invalid' })
+      }
+      else {
+        await Blog.findByIdAndDelete(request.params.id)
+        response.status(204).end()
+      } 
+      }
     catch(exception) {
       next(exception)
     }
